rag-system/requirements_updated.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.0
aiofiles==24.1.0
pydantic==2.9.0
groq==0.31.0
openai==1.50.0
anthropic==0.34.0
httpx==0.27.0
sqlalchemy==2.0.35
flask==3.0.3
flask-cors==4.0.1
